feat: :sparkles: arrangements

+++ ProyectoS1_Julian/src/main/java/com/example/ProyectoS1_Julian/service/impl/UsuarioServiceImpl.java
@@ -4,6 +4,7 @@ import com.example.ProyectoS1_Julian.dto.request.UsuarioRequestDTO;
 import com.example.ProyectoS1_Julian.dto.response.UsuarioResponseDTO;
 import com.example.ProyectoS1_Julian.mapper.UsuarioMapper;
 import com.example.ProyectoS1_Julian.modelo.Usuario;
+import com.example.ProyectoS1_Julian.repository.BodegaRepository;
 import com.example.ProyectoS1_Julian.repository.UsuarioRepository;
 import com.example.ProyectoS1_Julian.service.UsuarioService;
 import jakarta.persistence.EntityNotFoundException;
@@ -15,8 +16,10 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class UsuarioServiceImpl implements UsuarioService {
+
     private final UsuarioRepository usuarioRepository;
     private final UsuarioMapper usuarioMapper;
+    private final BodegaRepository bodegaRepository;
 
     @Override
     public UsuarioResponseDTO crear(UsuarioRequestDTO dto) {
@@ -54,6 +57,10 @@ public class UsuarioServiceImpl implements UsuarioService {
     public void eliminar(Long id) {
         if (!usuarioRepository.existsById(id)) {
             throw new EntityNotFoundException("Error: no existe el Usuario a eliminar");
+        }
+
+        if (!bodegaRepository.existsByUsuarioId(id)){
+            throw new EntityNotFoundException("Error, el usuario pertenece a una bodega");
         }
         usuarioRepository.deleteById(id);
     }
